axcient/axcient-utils: import encrypted pools from separate cachefile
The only way to allow unencrypted children of encrypted datasets, using
inherited mountpoints, is to load the encryption key at import time.
Otherwise the encrypted dataset may end up mounted on top of the unencrypted
one.  So we need to move the pool's cachefile, just like we did with geli,
to prevent it from being imported before Salt is ready.

CES-212