sysutils/smartmontools: shut up about USB bridges
If a typical user tries to run smartctl on a USB thumb drive, that indicates
that he's doing something wrong, and the cron job should rightly alert him.
But Axcient doesn't want to customize the cron job's settings on a
per-computer basis, and we don't use any USB-attached storage for
production, so such warnings are purely noise for us.  Silence them.

Do not upstream this change.