fix Axcient's custom golang -dev ports after 998a548dca0
These ports began to fail when pre-build tried to write to /nonexistent.
The cause was either the switch to go-1.19 or upstream 998a548dca0, which
changed the environment during the "make fetch" stage.  I'm not sure
which, but copying upstream's change from that revision into our
pre-build step fixes the problem.