Use dummy value when year/week is too large.
When converting from "unsigned char" to "char", sometimes bytes in the
week (and year) are of value 255, which is too big for signed char. When
that case is detected, use dummy value of -1 for year and week. This is
for now, until we discover how to properly read values from drives that
behave in this way.