Stop building net/openldap24-client
I'm unable to determine why we were building it in the first place.
openldap26-client is included as a dependency of something else we need,
but AFAIK we don't need openldap24-client for anything.  Also, that port
has been removed and replaced by openldap25-client, which we also don't
need.