Stop building emulators/open-vm-tools
The latest version of the port no longer builds in a FreeBSD 13.1 jail,
and we can't update our build jail to 13.2 because we still have servers
running 13.1.  This package was originally requested for use on VMWare
virtual machines.  But it isn't currently installed on any that I can
find, so there should be no harm in removing it.  If needed in the
future, the easiest way to build it again would be to either:
* Upgrade the last of our 13.1 servers to 13.2 or 14.0, or
* Upgrade all of the VMWare VMs to 14.0, and only build this port for
  FreeBSD 14.