axcient/szs-migrator: reduce false alarms in 900.szs-audit.in
There was a race condition.  Because szs-dataset-info is so slow, it was
possible for a dataset to finish migration after the initial "zfs list"
but before its "szs-dataset-info" ran.  If that happened, then it would
be reported as "misplaced" even though it was in fact no longer present.
The majority of our "misplaced dataset" alerts are of this type.

Fix the race condition by checking for the dataset's presence again
after "szs-dataset-info" runs.