net/ipxe: add custom patch (#974)
iPXE embeds your .ipxe scripts into the binaries during compilation.
Which makes it hard to make small adjustments after the bootloaders have
been built. Because of this we use this small script that starts dhcp
and then chain loads another script called "menu.ipxe". That way we can
do changes whenever we want.

menu.ipxe is just a normal text file that just needs to reside in the
same directory as the bootloader you're using.

The classic way to embed scripts is with
IPXE_MAKE_ARGS= EMBED=path_to_script. But this felt more appropriate.